Reworked controller implementation due to code review comments.
