+++ .gitignore
@@ -1,5 +1,5 @@
 # Project exclude paths
-/.gradle/
-/gradle/
-/.idea/
-/build/
+.gradle
+gradle
+.idea
+build
