+++ src/main/java/com/example/app/controller/UserController.java
@@ -3,25 +3,43 @@ package com.example.app.controller;
 import java.util.List;
 
 import org.springframework.web.bind.annotation.*;
+import org.springframework.beans.factory.annotation.Autowired;
 
+import com.example.app.service.UserService;
 import com.example.app.entity.User;
 
 /**
  * Controller for the {@link com.example.app.entity.User} entity.
  */
 
+@RestController
 @RequestMapping("/users")
-public interface UserController {
+public class UserController {
+
+    private final UserService service;
+
+    @Autowired
+    public UserController(UserService service) {
+        this.service = service;
+    }
 
     @GetMapping
-    List<User> findAll();
+    public List<User> findAll() {
+        return service.findAll();
+    }
 
     @GetMapping("/{id}")
-    User findById(@PathVariable("id") Long id);
+    public User findById(@PathVariable("id") Long id) {
+        return service.findById(id);
+    }
 
     @PostMapping
-    void save(@RequestBody User user);
+    public void save(@RequestBody User user) {
+        service.save(user);
+    }
 
     @DeleteMapping("/{id}")
-    void deleteById(@PathVariable("id") Long id);
+    public void deleteById(@PathVariable("id") Long id) {
+        service.deleteById(id);
+    }
 }
